fix: move state / usb go files to module wally
